move menu model -> constant
