+++ src/modules/home/components/Navbar/SideMenu/index.tsx
@@ -6,7 +6,7 @@ import {
 import React from 'react';
 import pathJoin from 'src/common/utils/path-join';
 import Link from 'next/link';
-import { APP_NAVBAR_CONSTANT } from 'Models/menu/constant';
+import { APP_NAVBAR_CONSTANT } from 'Modules/home/constants/menu';
 import { useRouter } from 'next/router';
 
 export type SideMenuProps = {
