+++ src/modules/home/constants/menu/index.tsx
@@ -1,21 +1,6 @@
 import Button from 'Components/Button';
 import React from 'react';
-import { MenuType } from './type';
-
-export const APP_SIDEBAR_CONSTANT: MenuType[] = [
-  {
-    key: '1',
-    label: 'Sign up',
-    link: ['register'],
-    active: false,
-  },
-  {
-    key: '2',
-    label: 'Login',
-    link: ['login'],
-    active: false,
-  },
-];
+import { MenuType } from './types';
 
 export const APP_NAVBAR_CONSTANT: MenuType[] = [
   {
@@ -31,6 +16,21 @@ export const APP_NAVBAR_CONSTANT: MenuType[] = [
         Login
       </Button>
     ),
+    link: ['login'],
+    active: false,
+  },
+];
+
+export const APP_SIDEBAR_CONSTANT: MenuType[] = [
+  {
+    key: '1',
+    label: 'Sign up',
+    link: ['register'],
+    active: false,
+  },
+  {
+    key: '2',
+    label: 'Login',
     link: ['login'],
     active: false,
   },
