+++ src/modules/home/constants/menu/types.ts
